+++ example/gems.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2025-2026, by Samuel Williams.
+
+source "https://rubygems.org"
+
+gem "bake"
+gem "falcon"
+
+gem "async-grpc", path: ".."
+gem "grpc"
+gem "grpc-tools"